backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
psycopg
psycopg-pool
requests
httpx
langchain
langchain-core
langchain-community
langchain-openai
langchain-anthropic
langchain-google-genai
langgraph
langgraph-checkpoint-postgres
openai